Prepare native and common for python runtime.
