+++ Code/Tools/TestImpactFramework/Frontend/Console/Native/Code/Source/TestImpactNativeRuntimeConfigurationFactory.h
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
+
+#include <TestImpactFramework/Native/TestImpactNativeConfiguration.h>
+
+namespace TestImpact
+{
+    //! Parses the native-soecific configuration data (in JSON format) and returns the constructed runtime configuration.
+    NativeRuntimeConfig NativeRuntimeConfigurationFactory(const AZStd::string& configurationData);
+} // namespace TestImpact
