+++ cmake/TestImpactFramework/LYTestImpactFramework.cmake
@@ -10,13 +10,13 @@
 set(LY_TEST_IMPACT_INSTRUMENTATION_BIN "" CACHE PATH "Path to test impact framework instrumentation binary")
 
 # Name of test impact framework console static library target
-set(LY_TEST_IMPACT_CONSOLE_STATIC_TARGET "TestImpact.Frontend.Console.Static")
+set(LY_TEST_IMPACT_CONSOLE_STATIC_TARGET "TestImpact.Frontend.Console.Native.Static")
 
 # Name of test impact framework python coverage gem target
 set(LY_TEST_IMPACT_PYTHON_COVERAGE_STATIC_TARGET "PythonCoverage.Editor.Static")
 
 # Name of test impact framework console target
-set(LY_TEST_IMPACT_CONSOLE_TARGET "TestImpact.Frontend.Console")
+set(LY_TEST_IMPACT_CONSOLE_TARGET "TestImpact.Frontend.Console.Native")
 
 # Directory for test impact artifacts and data
 set(LY_TEST_IMPACT_WORKING_DIR "${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/TestImpactFramework")
